+++ src/main/java/de/samply/notification/smtp/EmailRecipientType.java
@@ -1,0 +1,9 @@
+package de.samply.notification.smtp;
+
+public enum EmailRecipientType {
+    SESSION_USER,
+    EMAIL_ANNOTATION,
+    PROJECT_MANAGER_ADMIN,
+    BRIDGEHEAD_ADMIN,
+    PROJECT_ALL
+}