+++ src/main/java/de/samply/user/roles/UserProjectRoles.java
@@ -1,4 +1,13 @@
 package de.samply.user.roles;
 
+import de.samply.utils.ProjectRolesUtils;
+
+import java.util.List;
+
 public class UserProjectRoles extends UserRoles<ProjectRole> {
+
+    public List<ProjectRole> getBridgeheadRolesOrderedInDescendentTime(String bridgehead) {
+        return ProjectRolesUtils.orderCollectionInDescendentTime(getBridgeheadRoles(bridgehead));
+    }
+
 }